refactor: LoginStudentInfo -> StudentInfo로 변경 (#6)
