+++ src/main/java/com/team/buddyya/auth/domain/StudentInfo.java
@@ -1,16 +1,16 @@
-package com.team.buddyya.auth.dto;
+package com.team.buddyya.auth.domain;
 
 import com.team.buddyya.student.domain.Role;
 import com.team.buddyya.student.domain.Student;
 
-public record LoginStudentInfo(
+public record StudentInfo(
         Long id,
         String name,
         Role role
 ) {
 
-    public static LoginStudentInfo from(Student student) {
-        return new LoginStudentInfo(
+    public static StudentInfo from(Student student) {
+        return new StudentInfo(
                 student.getId(),
                 student.getName(),
                 student.getRole()
